HOTfixes breaking world generation (Multiverse-Core) [SD-8519]
These are changes not meant for production. Please squash this commit as soon as an actual fix has been applied... I shouldn't even create such a commit...

Anyway:
Multiverse-Core is too forceful when applying it's own configurations. We have to load on `STARTUP` to allow MVC to find our plugin and generator.
But we can't generate worlds this early in server boot... That's why we are listening for the event before doing any actual initialization... That's not good... Or at least it can't stay like this...